// packages/fans-society-contracts/types/truffle/contracts/AMM.d.ts
export interface Committed {
  name: "Committed";
  args: {
    id: BN;
    caller: string;
    amount: BN;
    0: BN;
    1: string;
    2: BN;
  };
}

export interface OwnershipTransferred {
  name: "OwnershipTransferred";
  args: {
    previousOwner: string;
    newOwner: string;
    0: string;
    1: string;
  };
}

export interface ProjectCreated {
  name: "ProjectCreated";
  args: {
    id: BN;
    name: string;
    symbol: string;
    description: string;
    target: BN;
    minInvest: BN;
    maxInvest: BN;
    totalSupply: BN;
    authorAddress: string;
    0: BN;
    1: string;
    2: string;
    3: string;
    4: BN;
    5: BN;
    6: BN;
    7: BN;
    8: string;
  };
}

export interface ProjectStatusChanged {
  name: "ProjectStatusChanged";
  args: {
    id: BN;
    status: BN;
    0: BN;
    1: BN;
  };
}

export interface TokensClaimed {
  name: "TokensClaimed";
  args: {
    projectId: BN;
    token: string;
    caller: string;
    amount: BN;
    0: BN;
    1: string;
    2: string;
    3: BN;
  };
}

export interface Withdrawed {
  name: "Withdrawed";
  args: {
    id: BN;
    caller: string;
    amount: BN;
    0: BN;
    1: string;
    2: BN;
  };
}

type AllEvents =
  | Committed
  | OwnershipTransferred
  | ProjectCreated
  | ProjectStatusChanged
  | TokensClaimed
  | Withdrawed;